Fix:#142 - delete carrer pagination comp and connect pagination
