+++ src/app/normal/my-page/written-review/page.tsx
@@ -2,7 +2,6 @@ import { HydrationBoundary, QueryClient, dehydrate } from '@tanstack/react-query
 import { getMyReviewData } from '@/api/ReviewService';
 import ReviewTabs from '@/components/Tabs/ReviewTabs';
 import NormalReviewCard from '@/components/cards/NormalReviewCard';
-import ReviewPagination from '@/pages/ReviewPagination';
 
 export default async function WrittenReviewPage() {
   const queryClient = new QueryClient();
@@ -24,7 +23,6 @@ export default async function WrittenReviewPage() {
           <div className="lg:grid lg:grid-cols-2 lg:gap-y-12 lg:gap-x-6 md:flex md:flex-col sm:flex sm:flex-col md:gap-y-8 sm:gap-y-8">
             <NormalReviewCard type="MY" />
           </div>
-          <ReviewPagination />
         </div>
       </>
     </HydrationBoundary>
