+++ src/interfaces/Modal/WritingReveiwModalInterface.ts
@@ -1,22 +1,6 @@
-import type { MyReviews } from '../Card/NormalReviewCardInterface';
+import type { MyReviews, ReviewCardEstimations } from '../Card/NormalReviewCardInterface';
 
 export interface WritingReviewModalProps {
-  estimation: {
-    id: string;
-    createdAt: string;
-    comment: string;
-    score: number;
-    driver: {
-      name: string;
-      image: string;
-    };
-    owner: {
-      moveInfos: {
-        type: 'SMALL' | 'HOME' | 'OFFICE';
-        date: string;
-        price: number;
-      };
-    };
-  };
+  estimation: ReviewCardEstimations;
   setIsModalOpen: (args: boolean) => void;
 }
